+++ src/main/kotlin/io/github/surpsg/DiffCoverageConfiguration.kt
@@ -1,4 +1,4 @@
-package com.github.surpsg
+package io.github.surpsg
 
 import java.io.File
 import java.net.URL